issue #133 - refactoring + stub for delegated harvesting
network currency and network harvesting mosaics sorted out a bit
